refactor(auth): split auth dto into separate files for better maintainability
Move auth-related DTOs from single file to individual files organized by purpose.
Update imports across the codebase to use new file structure.
Add new enums for auth and code types to improve type safety.
