+++ libs/security/index.ts
@@ -1,4 +1,5 @@
 export * from './jwt/jwt-auth.guard';
 export * from './jwt/jwt.strategy';
 export * from './decorators/public.decorator';
+export * from './decorators/user.decorator';
 export * from './types';
