+++ src/auth/enums/auth-type.enum.ts
@@ -1,0 +1,8 @@
+export enum AuthType {
+  USERNAME_PASSWORD = 'username_password',
+  EMAIL_PASSWORD = 'email_password',
+  EMAIL_CODE = 'email_code',
+  PHONE_PASSWORD = 'phone_password',
+  PHONE_CODE = 'phone_code',
+}
+
