+++ src/auth/enums/code-type.enum.ts
@@ -1,0 +1,6 @@
+export enum CodeType {
+  REGISTER = 'register',
+  LOGIN = 'login',
+  RESET_PASSWORD = 'reset_password',
+}
+
