+++ src/auth/repositories/verification.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../../prisma.service';
+import { PrismaService } from '@/prisma.service';
 import { VerificationCodeType } from '@prisma/client';
 
 @Injectable()
